CV/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121
torch 
torchvision 
torchaudio
wandb
PyYAML
hydra-core
tqdm
pandas
timm